benchmark/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.58.2
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.3
black==25.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
datasets==3.5.0
decorator==5.2.1
dill==0.3.8
distro==1.9.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2024.12.0
future==1.0.0
geocoder==1.38.1
google-auth==2.38.0
google-genai==1.7.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
jiter==0.9.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.9.4
mdurl==0.1.2
module-wrapper==0.3.1
multidict==6.4.3
multiprocess==0.70.16
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.106.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
propcache==0.3.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
ratelim==0.1.6
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.5.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stdlib-list==0.8.0
tabulate==0.9.0
together==1.5.21
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
websockets==15.0.1
wikipedia==1.4.0
xxhash==3.5.0
yarl==1.20.0
